Spring cleaning. Also a sweet rounded button class.
