+++ aw/ViewControllers/News/NewsDonateTableViewCell.swift
@@ -1,0 +1,24 @@
+//
+//  NewsDonateTableViewCell.swift
+//  aw
+//
+//  Created by Alex Kerney on 4/1/18.
+//  Copyright © 2018 Alex Kerney. All rights reserved.
+//
+
+import UIKit
+
+class NewsDonateTableViewCell: UITableViewCell {
+    @IBOutlet weak var donateButton: UIView!
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+
+}
