+++ aw/ViewControllers/Reaches/MapViewController.swift
@@ -65,7 +65,8 @@ extension MapViewController {
 
             do {
                 let reaches = try context.fetch(request)
-                if let reach = reaches.first, let detailVC = segue.destination as?   ReachDetailContainerViewController {
+                if let reach = reaches.first,
+                        let detailVC = segue.destination as?   ReachDetailContainerViewController {
                     detailVC.reach = reach
                 } else {
                     print("No reach")
